Improve tag name validation
Various pieces of the tagging system will break if the tag name contains
slashes or commas, so that should be checked for.

MediaWiki core (as of Ic5d75432) contains a method for doing just that,
so use it if possible.

Also require that tag names be either:
* Intended for reuse (canAddTagsAccompanyingChange)
* Already in use
* Creatable

This prevents the use of software-defined tags (unless whitelisted)
or tags that are blocked from creation (e.g. OAuth reserved tag
prefixes)

Add caching for tag lookup since it's used more often now.

Bug: T92956
Bug: T27151
Depends-On: Ic5d754323cbfd2c2b54c4df1245767946ebb1821
Change-Id: Ibc3095879f276a5436362f471b885e64f338522e